Move frontMatter style to line decos
